sh2: Refactor code into SH2 base CPU class + separate CPU devices
