+++ scripts/src/cpu.lua
@@ -853,32 +853,36 @@ end
 
 if CPUS["SH"] then
 	files {
+		MAME_DIR .. "src/devices/cpu/sh/sh_fe.cpp",
 		MAME_DIR .. "src/devices/cpu/sh/sh.cpp",
 		MAME_DIR .. "src/devices/cpu/sh/sh.h",
 		MAME_DIR .. "src/devices/cpu/sh/sh2.cpp",
 		MAME_DIR .. "src/devices/cpu/sh/sh2.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh2comn.cpp",
-		MAME_DIR .. "src/devices/cpu/sh/sh2comn.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh_fe.cpp",
 		MAME_DIR .. "src/devices/cpu/sh/sh2fe.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh3comn.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh3comn.h",
+		MAME_DIR .. "src/devices/cpu/sh/sh4.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh4.h",
+		MAME_DIR .. "src/devices/cpu/sh/sh4comn.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh4comn.h",
+		MAME_DIR .. "src/devices/cpu/sh/sh4dmac.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh4dmac.h",
 		MAME_DIR .. "src/devices/cpu/sh/sh4fe.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh4regs.h",
+		MAME_DIR .. "src/devices/cpu/sh/sh4tmu.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh4tmu.h",
+		MAME_DIR .. "src/devices/cpu/sh/sh7021.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh7021.h",
+		MAME_DIR .. "src/devices/cpu/sh/sh7032.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh7032.h",
 		MAME_DIR .. "src/devices/cpu/sh/sh7604_bus.cpp",
 		MAME_DIR .. "src/devices/cpu/sh/sh7604_bus.h",
 		MAME_DIR .. "src/devices/cpu/sh/sh7604_sci.cpp",
 		MAME_DIR .. "src/devices/cpu/sh/sh7604_sci.h",
 		MAME_DIR .. "src/devices/cpu/sh/sh7604_wdt.cpp",
 		MAME_DIR .. "src/devices/cpu/sh/sh7604_wdt.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh4.cpp",
-		MAME_DIR .. "src/devices/cpu/sh/sh4.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh4comn.cpp",
-		MAME_DIR .. "src/devices/cpu/sh/sh4comn.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh3comn.cpp",
-		MAME_DIR .. "src/devices/cpu/sh/sh3comn.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh4tmu.cpp",
-		MAME_DIR .. "src/devices/cpu/sh/sh4tmu.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh4dmac.cpp",
-		MAME_DIR .. "src/devices/cpu/sh/sh4dmac.h",
-		MAME_DIR .. "src/devices/cpu/sh/sh4regs.h",
+		MAME_DIR .. "src/devices/cpu/sh/sh7604.cpp",
+		MAME_DIR .. "src/devices/cpu/sh/sh7604.h",
 	}
 end
