+++ src/devices/cpu/sh/sh7604.cpp
@@ -1,25 +1,310 @@
 // license:BSD-3-Clause
-// copyright-holders:R. Belmont
+// copyright-holders:Juergen Buchmueller, R. Belmont, Angelo Salese, ...
 /*****************************************************************************
  *
- *   sh2common.c
+ *   sh2.c
+ *   Portable Hitachi SH-2 (SH7600 family) emulator
  *
- *   SH-2 non-specific components
+ *  This work is based on <tiraniddo@hotmail.com> C/C++ implementation of
+ *  the SH-2 CPU core and was adapted to the MAME CPU core requirements.
+ *  Thanks also go to Chuck Mason <chukjr@sundail.net> and Olivier Galibert
+ *  <galibert@pobox.com> for letting me peek into their SEMU code :-)
  *
  *****************************************************************************/
+/*
+TODO: Test and use sh7604_wdt_device, sh7604_sci_device, and sh7604_bus_device as appropriate
+*/
+
 
 #include "emu.h"
-#include "sh2.h"
-#include "sh2comn.h"
+
+#include "sh7604.h"
 
 //#define VERBOSE 1
 #include "logmacro.h"
 
-static const int div_tab[4] = { 3, 5, 7, 0 };
-static const int wdtclk_tab[8] = { 1, 6, 7, 8, 9, 10, 12, 13 };
+
+static constexpr int div_tab[4] = { 3, 5, 7, 0 };
+static constexpr int wdtclk_tab[8] = { 1, 6, 7, 8, 9, 10, 12, 13 };
 
 
-void sh2_device::sh2_timer_resync()
+DEFINE_DEVICE_TYPE(SH2_SH7604,  sh2_sh7604_device,  "sh2_7604",  "Hitachi SH-2 (SH7604)")
+
+
+sh2_sh7604_device::sh2_sh7604_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock)
+	: sh2_device(mconfig, SH2_SH7604, tag, owner, clock, CPU_TYPE_SH2, address_map_constructor(FUNC(sh2_sh7604_device::sh7604_map), this), 32, 0xc7ffffff)
+	, m_is_slave(0)
+	, m_dma_kludge_cb(*this)
+	, m_dma_fifo_data_available_cb(*this)
+	, m_ftcsr_read_cb(*this)
+{
+}
+
+void sh2_sh7604_device::device_start()
+{
+	sh2_device::device_start();
+
+	m_timer = timer_alloc(FUNC(sh2_sh7604_device::sh2_timer_callback), this);
+	m_timer->adjust(attotime::never);
+	m_wdtimer = timer_alloc(FUNC(sh2_sh7604_device::sh2_wdtimer_callback), this);
+	m_wdtimer->adjust(attotime::never);
+
+	m_dma_current_active_timer[0] = timer_alloc(FUNC(sh2_sh7604_device::sh2_dma_current_active_callback), this);
+	m_dma_current_active_timer[0]->adjust(attotime::never);
+
+	m_dma_current_active_timer[1] = timer_alloc(FUNC(sh2_sh7604_device::sh2_dma_current_active_callback), this);
+	m_dma_current_active_timer[1]->adjust(attotime::never);
+
+	/* resolve callbacks */
+	m_dma_kludge_cb.resolve();
+	m_dma_fifo_data_available_cb.resolve();
+	m_ftcsr_read_cb.resolve();
+
+	// SCI
+	save_item(NAME(m_smr));
+	save_item(NAME(m_brr));
+	save_item(NAME(m_scr));
+	save_item(NAME(m_tdr));
+	save_item(NAME(m_ssr));
+
+	// FRT / FRC
+	save_item(NAME(m_tier));
+	save_item(NAME(m_ftcsr));
+	save_item(NAME(m_frc_tcr));
+	save_item(NAME(m_tocr));
+	save_item(NAME(m_frc));
+	save_item(NAME(m_ocra));
+	save_item(NAME(m_ocrb));
+	save_item(NAME(m_frc_icr));
+	save_item(NAME(m_frc_base));
+	save_item(NAME(m_frt_input));
+
+	// INTC
+	save_item(NAME(m_irq_level.frc));
+	save_item(NAME(m_irq_level.sci));
+	save_item(NAME(m_irq_level.divu));
+	save_item(NAME(m_irq_level.dmac));
+	save_item(NAME(m_irq_level.wdt));
+	save_item(NAME(m_irq_vector.fic));
+	save_item(NAME(m_irq_vector.foc));
+	save_item(NAME(m_irq_vector.fov));
+	save_item(NAME(m_irq_vector.divu));
+	save_item(NAME(m_irq_vector.dmac));
+
+	save_item(NAME(m_ipra));
+	save_item(NAME(m_iprb));
+	save_item(NAME(m_vcra));
+	save_item(NAME(m_vcrb));
+	save_item(NAME(m_vcrc));
+	save_item(NAME(m_vcrd));
+	save_item(NAME(m_vcrwdt));
+	save_item(NAME(m_vcrdiv));
+	save_item(NAME(m_intc_icr));
+	save_item(NAME(m_vcrdma));
+
+	save_item(NAME(m_vecmd));
+	save_item(NAME(m_nmie));
+
+	// DIVU
+	save_item(NAME(m_divu_ovf));
+	save_item(NAME(m_divu_ovfie));
+	save_item(NAME(m_dvsr));
+	save_item(NAME(m_dvdntl));
+	save_item(NAME(m_dvdnth));
+
+	// WTC
+	save_item(NAME(m_wtcnt));
+	save_item(NAME(m_wtcsr));
+	save_item(NAME(m_rstcsr));
+	save_item(NAME(m_wtcw));
+
+	// DMAC
+	save_item(NAME(m_dmaor));
+	save_item(STRUCT_MEMBER(m_dmac, drcr));
+	save_item(STRUCT_MEMBER(m_dmac, sar));
+	save_item(STRUCT_MEMBER(m_dmac, dar));
+	save_item(STRUCT_MEMBER(m_dmac, tcr));
+	save_item(STRUCT_MEMBER(m_dmac, chcr));
+
+	// misc
+	save_item(NAME(m_sbycr));
+	save_item(NAME(m_ccr));
+
+	// BSC
+	save_item(NAME(m_bcr1));
+	save_item(NAME(m_bcr2));
+	save_item(NAME(m_wcr));
+	save_item(NAME(m_mcr));
+	save_item(NAME(m_rtcsr));
+	save_item(NAME(m_rtcor));
+	save_item(NAME(m_rtcnt));
+}
+
+void sh2_sh7604_device::device_reset()
+{
+	sh2_device::device_reset();
+
+	m_frc = 0;
+	m_ocra = 0;
+	m_ocrb = 0;
+	m_frc_icr = 0;
+	m_frc_base = 0;
+	m_frt_input = 0;
+
+	for ( int i = 0; i < 2; i++ )
+	{
+		m_dma_timer_active[i] = 0;
+		m_dma_irq[i] = 0;
+		m_active_dma_incs[i] = 0;
+		m_active_dma_incd[i] = 0;
+		m_active_dma_size[i] = 0;
+		m_active_dma_steal[i] = 0;
+		m_active_dma_src[i] = 0;
+		m_active_dma_dst[i] = 0;
+		m_active_dma_count[i] = 0;
+	}
+
+	m_wtcnt = 0;
+	m_wtcsr = 0;
+}
+
+void sh2_sh7604_device::sh7604_map(address_map &map)
+{
+	map(0x40000000, 0xbfffffff).r(FUNC(sh2_sh7604_device::sh2_internal_a5));
+
+//  TODO: cps3boot breaks with this enabled. Needs callback
+//  map(0xc0000000, 0xc0000fff).ram(); // cache data array
+
+//  map(0xe0000000, 0xe00001ff).mirror(0x1ffffe00).rw(FUNC(sh2_sh7604_device::sh7604_r), FUNC(sh2_sh7604_device::sh7604_w));
+	// TODO: internal map takes way too much resources if mirrored with 0x1ffffe00
+	//       we eventually internalize again via trampoline & sh7604_device
+	//       Also area 0xffff8000-0xffffbfff is for synchronous DRAM mode,
+	//       so this isn't actually a full mirror
+	// SCI
+	map(0xfffffe00, 0xfffffe00).rw(FUNC(sh2_sh7604_device::smr_r), FUNC(sh2_sh7604_device::smr_w));
+	map(0xfffffe01, 0xfffffe01).rw(FUNC(sh2_sh7604_device::brr_r), FUNC(sh2_sh7604_device::brr_w));
+	map(0xfffffe02, 0xfffffe02).rw(FUNC(sh2_sh7604_device::scr_r), FUNC(sh2_sh7604_device::scr_w));
+	map(0xfffffe03, 0xfffffe03).rw(FUNC(sh2_sh7604_device::tdr_r), FUNC(sh2_sh7604_device::tdr_w));
+	map(0xfffffe04, 0xfffffe04).rw(FUNC(sh2_sh7604_device::ssr_r), FUNC(sh2_sh7604_device::ssr_w));
+	map(0xfffffe05, 0xfffffe05).r(FUNC(sh2_sh7604_device::rdr_r));
+
+	// FRC
+	map(0xfffffe10, 0xfffffe10).rw(FUNC(sh2_sh7604_device::tier_r), FUNC(sh2_sh7604_device::tier_w));
+	map(0xfffffe11, 0xfffffe11).rw(FUNC(sh2_sh7604_device::ftcsr_r), FUNC(sh2_sh7604_device::ftcsr_w));
+	map(0xfffffe12, 0xfffffe13).rw(FUNC(sh2_sh7604_device::frc_r), FUNC(sh2_sh7604_device::frc_w));
+	map(0xfffffe14, 0xfffffe15).rw(FUNC(sh2_sh7604_device::ocra_b_r), FUNC(sh2_sh7604_device::ocra_b_w));
+	map(0xfffffe16, 0xfffffe16).rw(FUNC(sh2_sh7604_device::frc_tcr_r), FUNC(sh2_sh7604_device::frc_tcr_w));
+	map(0xfffffe17, 0xfffffe17).rw(FUNC(sh2_sh7604_device::tocr_r), FUNC(sh2_sh7604_device::tocr_w));
+	map(0xfffffe18, 0xfffffe19).r(FUNC(sh2_sh7604_device::frc_icr_r));
+
+	// INTC
+	map(0xfffffe60, 0xfffffe61).rw(FUNC(sh2_sh7604_device::iprb_r), FUNC(sh2_sh7604_device::iprb_w));
+	map(0xfffffe62, 0xfffffe63).rw(FUNC(sh2_sh7604_device::vcra_r), FUNC(sh2_sh7604_device::vcra_w));
+	map(0xfffffe64, 0xfffffe65).rw(FUNC(sh2_sh7604_device::vcrb_r), FUNC(sh2_sh7604_device::vcrb_w));
+	map(0xfffffe66, 0xfffffe67).rw(FUNC(sh2_sh7604_device::vcrc_r), FUNC(sh2_sh7604_device::vcrc_w));
+	map(0xfffffe68, 0xfffffe69).rw(FUNC(sh2_sh7604_device::vcrd_r), FUNC(sh2_sh7604_device::vcrd_w));
+
+	map(0xfffffe71, 0xfffffe71).rw(FUNC(sh2_sh7604_device::drcr_r<0>), FUNC(sh2_sh7604_device::drcr_w<0>));
+	map(0xfffffe72, 0xfffffe72).rw(FUNC(sh2_sh7604_device::drcr_r<1>), FUNC(sh2_sh7604_device::drcr_w<1>));
+
+	// WTC
+	map(0xfffffe80, 0xfffffe81).rw(FUNC(sh2_sh7604_device::wtcnt_r), FUNC(sh2_sh7604_device::wtcnt_w));
+	map(0xfffffe82, 0xfffffe83).rw(FUNC(sh2_sh7604_device::rstcsr_r), FUNC(sh2_sh7604_device::rstcsr_w));
+
+	// standby and cache control
+	map(0xfffffe90, 0xfffffe91).rw(FUNC(sh2_sh7604_device::fmr_sbycr_r), FUNC(sh2_sh7604_device::fmr_sbycr_w));
+	map(0xfffffe92, 0xfffffe92).rw(FUNC(sh2_sh7604_device::ccr_r), FUNC(sh2_sh7604_device::ccr_w));
+
+	// INTC second section
+	map(0xfffffee0, 0xfffffee1).rw(FUNC(sh2_sh7604_device::intc_icr_r), FUNC(sh2_sh7604_device::intc_icr_w));
+	map(0xfffffee2, 0xfffffee3).rw(FUNC(sh2_sh7604_device::ipra_r), FUNC(sh2_sh7604_device::ipra_w));
+	map(0xfffffee4, 0xfffffee5).rw(FUNC(sh2_sh7604_device::vcrwdt_r), FUNC(sh2_sh7604_device::vcrwdt_w));
+
+	// DIVU
+	map(0xffffff00, 0xffffff03).rw(FUNC(sh2_sh7604_device::dvsr_r), FUNC(sh2_sh7604_device::dvsr_w));
+	map(0xffffff04, 0xffffff07).rw(FUNC(sh2_sh7604_device::dvdnt_r), FUNC(sh2_sh7604_device::dvdnt_w));
+	map(0xffffff08, 0xffffff0b).rw(FUNC(sh2_sh7604_device::dvcr_r), FUNC(sh2_sh7604_device::dvcr_w));
+	// INTC third section
+	map(0xffffff0c, 0xffffff0f).rw(FUNC(sh2_sh7604_device::vcrdiv_r), FUNC(sh2_sh7604_device::vcrdiv_w));
+	// DIVU continued (64-bit plus mirrors)
+	map(0xffffff10, 0xffffff13).rw(FUNC(sh2_sh7604_device::dvdnth_r), FUNC(sh2_sh7604_device::dvdnth_w));
+	map(0xffffff14, 0xffffff17).rw(FUNC(sh2_sh7604_device::dvdntl_r), FUNC(sh2_sh7604_device::dvdntl_w));
+	map(0xffffff18, 0xffffff1b).r(FUNC(sh2_sh7604_device::dvdnth_r));
+	map(0xffffff1c, 0xffffff1f).r(FUNC(sh2_sh7604_device::dvdntl_r));
+
+	// DMAC
+	map(0xffffff80, 0xffffff83).rw(FUNC(sh2_sh7604_device::sar_r<0>), FUNC(sh2_sh7604_device::sar_w<0>));
+	map(0xffffff84, 0xffffff87).rw(FUNC(sh2_sh7604_device::dar_r<0>), FUNC(sh2_sh7604_device::dar_w<0>));
+	map(0xffffff88, 0xffffff8b).rw(FUNC(sh2_sh7604_device::dmac_tcr_r<0>), FUNC(sh2_sh7604_device::dmac_tcr_w<0>));
+	map(0xffffff8c, 0xffffff8f).rw(FUNC(sh2_sh7604_device::chcr_r<0>), FUNC(sh2_sh7604_device::chcr_w<0>));
+
+	map(0xffffff90, 0xffffff93).rw(FUNC(sh2_sh7604_device::sar_r<1>), FUNC(sh2_sh7604_device::sar_w<1>));
+	map(0xffffff94, 0xffffff97).rw(FUNC(sh2_sh7604_device::dar_r<1>), FUNC(sh2_sh7604_device::dar_w<1>));
+	map(0xffffff98, 0xffffff9b).rw(FUNC(sh2_sh7604_device::dmac_tcr_r<1>), FUNC(sh2_sh7604_device::dmac_tcr_w<1>));
+	map(0xffffff9c, 0xffffff9f).rw(FUNC(sh2_sh7604_device::chcr_r<1>), FUNC(sh2_sh7604_device::chcr_w<1>));
+
+	map(0xffffffa0, 0xffffffa3).rw(FUNC(sh2_sh7604_device::vcrdma_r<0>), FUNC(sh2_sh7604_device::vcrdma_w<0>));
+	map(0xffffffa8, 0xffffffab).rw(FUNC(sh2_sh7604_device::vcrdma_r<1>), FUNC(sh2_sh7604_device::vcrdma_w<1>));
+	map(0xffffffb0, 0xffffffb3).rw(FUNC(sh2_sh7604_device::dmaor_r), FUNC(sh2_sh7604_device::dmaor_w));
+
+	// BSC
+	map(0xffffffe0, 0xffffffe3).rw(FUNC(sh2_sh7604_device::bcr1_r), FUNC(sh2_sh7604_device::bcr1_w));
+	map(0xffffffe4, 0xffffffe7).rw(FUNC(sh2_sh7604_device::bcr2_r), FUNC(sh2_sh7604_device::bcr2_w));
+	map(0xffffffe8, 0xffffffeb).rw(FUNC(sh2_sh7604_device::wcr_r), FUNC(sh2_sh7604_device::wcr_w));
+	map(0xffffffec, 0xffffffef).rw(FUNC(sh2_sh7604_device::mcr_r), FUNC(sh2_sh7604_device::mcr_w));
+	map(0xfffffff0, 0xfffffff3).rw(FUNC(sh2_sh7604_device::rtcsr_r), FUNC(sh2_sh7604_device::rtcsr_w));
+	map(0xfffffff4, 0xfffffff7).rw(FUNC(sh2_sh7604_device::rtcnt_r), FUNC(sh2_sh7604_device::rtcnt_w));
+	map(0xfffffff8, 0xfffffffb).rw(FUNC(sh2_sh7604_device::rtcor_r), FUNC(sh2_sh7604_device::rtcor_w));
+}
+
+
+void sh2_sh7604_device::sh2_exception(const char *message, int irqline)
+{
+	int vector;
+
+	if (irqline != 16)
+	{
+		if (irqline <= ((m_sh2_state->sr >> 4) & 15)) /* If the cpu forbids this interrupt */
+			return;
+
+		// if this is an sh2 internal irq, use its vector
+		if (m_sh2_state->internal_irq_level == irqline)
+		{
+			vector = m_internal_irq_vector;
+			/* avoid spurious irqs with this (TODO: needs a better fix) */
+			m_sh2_state->internal_irq_level = -1;
+			LOG("SH-2 exception #%d (internal vector: $%x) after [%s]\n", irqline, vector, message);
+		}
+		else
+		{
+			if(m_vecmd == true)
+			{
+				vector = standard_irq_callback(irqline, m_sh2_state->pc);
+				LOG("SH-2 exception #%d (external vector: $%x) after [%s]\n", irqline, vector, message);
+			}
+			else
+			{
+				standard_irq_callback(irqline, m_sh2_state->pc);
+				vector = 64 + irqline/2;
+				LOG("SH-2 exception #%d (autovector: $%x) after [%s]\n", irqline, vector, message);
+			}
+		}
+	}
+	else
+	{
+		vector = 11;
+		LOG("SH-2 nmi exception (autovector: $%x) after [%s]\n", vector, message);
+	}
+
+	sh2_exception_internal(message, irqline, vector);
+}
+
+uint32_t sh2_sh7604_device::sh2_internal_a5()
+{
+	return 0xa5a5a5a5;
+}
+
+void sh2_sh7604_device::sh2_timer_resync()
 {
 	// TODO: setting 3 is "External clock: count on rising edge"
 	int divider = div_tab[m_frc_tcr & 3];
@@ -35,7 +320,7 @@ void sh2_device::sh2_timer_resync()
 	}
 }
 
-void sh2_device::sh2_timer_activate()
+void sh2_sh7604_device::sh2_timer_activate()
 {
 	int max_delta = 0xfffff;
 	uint16_t frc;
@@ -73,7 +358,7 @@ void sh2_device::sh2_timer_activate()
 	}
 }
 
-TIMER_CALLBACK_MEMBER( sh2_device::sh2_timer_callback )
+TIMER_CALLBACK_MEMBER( sh2_sh7604_device::sh2_timer_callback )
 {
 	uint16_t frc;
 
@@ -99,18 +384,18 @@ TIMER_CALLBACK_MEMBER( sh2_device::sh2_timer_callback )
 	sh2_timer_activate();
 }
 
-void sh2_device::sh2_wtcnt_recalc()
+void sh2_sh7604_device::sh2_wtcnt_recalc()
 {
 	if (m_wdtimer->expire() != attotime::never)
 		m_wtcnt = 0x100 - (attotime_to_cycles(m_wdtimer->remaining()) >> wdtclk_tab[m_wtcsr & 7]);
 }
 
-void sh2_device::sh2_wdt_activate()
+void sh2_sh7604_device::sh2_wdt_activate()
 {
 	m_wdtimer->adjust(cycles_to_attotime((0x100 - m_wtcnt) << wdtclk_tab[m_wtcsr & 7]));
 }
 
-TIMER_CALLBACK_MEMBER(sh2_device::sh2_wdtimer_callback)
+TIMER_CALLBACK_MEMBER(sh2_sh7604_device::sh2_wdtimer_callback)
 {
 	m_wtcnt = 0;
 	if (!(m_wtcsr & 0x40))  // timer mode
@@ -150,7 +435,7 @@ TIMER_CALLBACK_MEMBER(sh2_device::sh2_wdtimer_callback)
 
 
 
-void sh2_device::sh2_notify_dma_data_available()
+void sh2_sh7604_device::sh2_notify_dma_data_available()
 {
 	//printf("call notify\n");
 
@@ -168,7 +453,7 @@ void sh2_device::sh2_notify_dma_data_available()
 
 }
 
-void sh2_device::sh2_do_dma(int dmach)
+void sh2_sh7604_device::sh2_do_dma(int dmach)
 {
 	uint32_t dmadata;
 
@@ -392,7 +677,7 @@ void sh2_device::sh2_do_dma(int dmach)
 	}
 }
 
-TIMER_CALLBACK_MEMBER( sh2_device::sh2_dma_current_active_callback )
+TIMER_CALLBACK_MEMBER( sh2_sh7604_device::sh2_dma_current_active_callback )
 {
 	int dma = param & 1;
 
@@ -400,7 +685,7 @@ TIMER_CALLBACK_MEMBER( sh2_device::sh2_dma_current_active_callback )
 }
 
 
-void sh2_device::sh2_dmac_check(int dmach)
+void sh2_sh7604_device::sh2_dmac_check(int dmach)
 {
 	if(m_dmac[dmach].chcr & m_dmaor & 1)
 	{
@@ -426,8 +711,8 @@ void sh2_device::sh2_dmac_check(int dmach)
 
 			m_dma_timer_active[dmach] = 1;
 
-			m_active_dma_src[dmach] &= SH12_AM;
-			m_active_dma_dst[dmach] &= SH12_AM;
+			m_active_dma_src[dmach] &= m_am;
+			m_active_dma_dst[dmach] &= m_am;
 
 			switch(m_active_dma_size[dmach])
 			{
@@ -478,59 +763,59 @@ void sh2_device::sh2_dmac_check(int dmach)
  */
 // TODO: identical to H8 counterpart
 
-uint8_t sh2_device::smr_r()
+uint8_t sh2_sh7604_device::smr_r()
 {
 	return m_smr;
 }
 
-void sh2_device::smr_w(uint8_t data)
+void sh2_sh7604_device::smr_w(uint8_t data)
 {
 	m_smr = data;
 }
 
-uint8_t sh2_device::brr_r()
+uint8_t sh2_sh7604_device::brr_r()
 {
 	return m_brr;
 }
 
-void sh2_device::brr_w(uint8_t data)
+void sh2_sh7604_device::brr_w(uint8_t data)
 {
 	m_brr = data;
 }
 
-uint8_t sh2_device::scr_r()
+uint8_t sh2_sh7604_device::scr_r()
 {
 	return m_scr;
 }
 
-void sh2_device::scr_w(uint8_t data)
+void sh2_sh7604_device::scr_w(uint8_t data)
 {
 	m_scr = data;
 }
 
-uint8_t sh2_device::tdr_r()
+uint8_t sh2_sh7604_device::tdr_r()
 {
 	return m_tdr;
 }
 
-void sh2_device::tdr_w(uint8_t data)
+void sh2_sh7604_device::tdr_w(uint8_t data)
 {
 	m_tdr = data;
 	// printf("%c",data & 0xff);
 }
 
-uint8_t sh2_device::ssr_r()
+uint8_t sh2_sh7604_device::ssr_r()
 {
 	// 0x84 is needed by EGWord on Saturn to make it to boot for some reason.
 	return m_ssr | 0x84;
 }
 
-void sh2_device::ssr_w(uint8_t data)
+void sh2_sh7604_device::ssr_w(uint8_t data)
 {
 	m_ssr = data;
 }
 
-uint8_t sh2_device::rdr_r()
+uint8_t sh2_sh7604_device::rdr_r()
 {
 	return 0;
 }
@@ -539,12 +824,12 @@ uint8_t sh2_device::rdr_r()
  * FRC
  */
 
-uint8_t sh2_device::tier_r()
+uint8_t sh2_sh7604_device::tier_r()
 {
 	return m_tier;
 }
 
-void sh2_device::tier_w(uint8_t data)
+void sh2_sh7604_device::tier_w(uint8_t data)
 {
 	sh2_timer_resync();
 	m_tier = data;
@@ -552,7 +837,7 @@ void sh2_device::tier_w(uint8_t data)
 	sh2_recalc_irq();
 }
 
-uint8_t sh2_device::ftcsr_r()
+uint8_t sh2_sh7604_device::ftcsr_r()
 {
 	// TODO: to be tested
 	if (!m_ftcsr_read_cb.isnull())
@@ -561,7 +846,7 @@ uint8_t sh2_device::ftcsr_r()
 	return m_ftcsr;
 }
 
-void sh2_device::ftcsr_w(uint8_t data)
+void sh2_sh7604_device::ftcsr_w(uint8_t data)
 {
 	uint8_t old;
 	old = m_ftcsr;
@@ -573,13 +858,13 @@ void sh2_device::ftcsr_w(uint8_t data)
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::frc_r()
+uint16_t sh2_sh7604_device::frc_r()
 {
 	sh2_timer_resync();
 	return m_frc;
 }
 
-void sh2_device::frc_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::frc_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	sh2_timer_resync();
 	COMBINE_DATA(&m_frc);
@@ -587,12 +872,12 @@ void sh2_device::frc_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::ocra_b_r()
+uint16_t sh2_sh7604_device::ocra_b_r()
 {
 	return (m_tocr & 0x10) ? m_ocrb : m_ocra;
 }
 
-void sh2_device::ocra_b_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::ocra_b_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	sh2_timer_resync();
 	if(m_tocr & 0x10)
@@ -603,12 +888,12 @@ void sh2_device::ocra_b_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	sh2_recalc_irq();
 }
 
-uint8_t sh2_device::frc_tcr_r()
+uint8_t sh2_sh7604_device::frc_tcr_r()
 {
 	return m_frc_tcr & 0x83;
 }
 
-void sh2_device::frc_tcr_w(uint8_t data)
+void sh2_sh7604_device::frc_tcr_w(uint8_t data)
 {
 	sh2_timer_resync();
 	m_frc_tcr = data & 0x83;
@@ -616,12 +901,12 @@ void sh2_device::frc_tcr_w(uint8_t data)
 	sh2_recalc_irq();
 }
 
-uint8_t sh2_device::tocr_r()
+uint8_t sh2_sh7604_device::tocr_r()
 {
 	return (m_tocr & 0x13) | 0xe0;
 }
 
-void sh2_device::tocr_w(uint8_t data)
+void sh2_sh7604_device::tocr_w(uint8_t data)
 {
 	sh2_timer_resync();
 	// TODO: output levels A/B (bits 1-0)
@@ -630,7 +915,7 @@ void sh2_device::tocr_w(uint8_t data)
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::frc_icr_r()
+uint16_t sh2_sh7604_device::frc_icr_r()
 {
 	return m_frc_icr;
 }
@@ -639,26 +924,26 @@ uint16_t sh2_device::frc_icr_r()
  * INTC
  */
 
-uint16_t sh2_device::intc_icr_r()
+uint16_t sh2_sh7604_device::intc_icr_r()
 {
 	// TODO: flip meaning based off NMI edge select bit (NMIE)
 	uint16_t nmilv = m_nmi_line_state == ASSERT_LINE ? 0 : 0x8000;
 	return (nmilv) | (m_intc_icr & 0x0101);
 }
 
-void sh2_device::intc_icr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::intc_icr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_intc_icr);
 	m_nmie = bool(BIT(m_intc_icr, 8));
 	m_vecmd = bool(BIT(m_intc_icr, 0));
 }
 
-uint16_t sh2_device::ipra_r()
+uint16_t sh2_sh7604_device::ipra_r()
 {
 	return m_ipra & 0xfff0;
 }
 
-void sh2_device::ipra_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::ipra_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_ipra);
 	m_irq_level.divu = (m_ipra >> 12) & 0xf;
@@ -667,12 +952,12 @@ void sh2_device::ipra_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::iprb_r()
+uint16_t sh2_sh7604_device::iprb_r()
 {
 	return m_iprb & 0xff00;
 }
 
-void sh2_device::iprb_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::iprb_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_iprb);
 	m_irq_level.sci = (m_iprb >> 12) & 0xf;
@@ -680,36 +965,36 @@ void sh2_device::iprb_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::vcra_r()
+uint16_t sh2_sh7604_device::vcra_r()
 {
 	return m_vcra & 0x7f7f;
 }
 
-void sh2_device::vcra_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::vcra_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_vcra);
 	// ...
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::vcrb_r()
+uint16_t sh2_sh7604_device::vcrb_r()
 {
 	return m_vcrb;
 }
 
-void sh2_device::vcrb_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::vcrb_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_vcrb);
 	// ...
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::vcrc_r()
+uint16_t sh2_sh7604_device::vcrc_r()
 {
 	return m_vcrc & 0x7f7f;
 }
 
-void sh2_device::vcrc_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::vcrc_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_vcrc);
 	m_irq_vector.fic = (m_vcrc >> 8) & 0x7f;
@@ -717,36 +1002,36 @@ void sh2_device::vcrc_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::vcrd_r()
+uint16_t sh2_sh7604_device::vcrd_r()
 {
 	return m_vcrd & 0x7f00;
 }
 
-void sh2_device::vcrd_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::vcrd_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_vcrd);
 	m_irq_vector.fov = (m_vcrc >> 8) & 0x7f;
 	sh2_recalc_irq();
 }
 
-uint16_t sh2_device::vcrwdt_r()
+uint16_t sh2_sh7604_device::vcrwdt_r()
 {
 	return m_vcrwdt & 0x7f7f;
 }
 
-void sh2_device::vcrwdt_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::vcrwdt_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_vcrwdt);
 	// ...
 	sh2_recalc_irq();
 }
 
-uint32_t sh2_device::vcrdiv_r()
+uint32_t sh2_sh7604_device::vcrdiv_r()
 {
 	return m_vcrdiv & 0x7f;
 }
 
-void sh2_device::vcrdiv_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::vcrdiv_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_vcrdiv);
 	// TODO: unemulated, level is seemingly not documented/settable?
@@ -758,12 +1043,12 @@ void sh2_device::vcrdiv_w(offs_t offset, uint32_t data, uint32_t mem_mask)
  * DIVU
  */
 
-uint32_t sh2_device::dvcr_r()
+uint32_t sh2_sh7604_device::dvcr_r()
 {
 	return (m_divu_ovfie == true ? 2 : 0) | (m_divu_ovf == true ? 1 : 0);
 }
 
-void sh2_device::dvcr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::dvcr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	if(ACCESSING_BITS_0_7)
 	{
@@ -779,22 +1064,22 @@ void sh2_device::dvcr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 	}
 }
 
-uint32_t sh2_device::dvsr_r()
+uint32_t sh2_sh7604_device::dvsr_r()
 {
 	return m_dvsr;
 }
 
-void sh2_device::dvsr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::dvsr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_dvsr);
 }
 
-uint32_t sh2_device::dvdnt_r()
+uint32_t sh2_sh7604_device::dvdnt_r()
 {
 	return m_dvdntl;
 }
 
-void sh2_device::dvdnt_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::dvdnt_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_dvdntl);
 	int32_t a = m_dvdntl;
@@ -814,22 +1099,22 @@ void sh2_device::dvdnt_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 	}
 }
 
-uint32_t sh2_device::dvdnth_r()
+uint32_t sh2_sh7604_device::dvdnth_r()
 {
 	return m_dvdnth;
 }
 
-uint32_t sh2_device::dvdntl_r()
+uint32_t sh2_sh7604_device::dvdntl_r()
 {
 	return m_dvdntl;
 }
 
-void sh2_device::dvdnth_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::dvdnth_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_dvdnth);
 }
 
-void sh2_device::dvdntl_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::dvdntl_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_dvdntl);
 	int64_t a = m_dvdntl | ((uint64_t)(m_dvdnth) << 32);
@@ -864,18 +1149,18 @@ void sh2_device::dvdntl_w(offs_t offset, uint32_t data, uint32_t mem_mask)
  * WTC
  */
 
-uint16_t sh2_device::wtcnt_r()
+uint16_t sh2_sh7604_device::wtcnt_r()
 {
 	sh2_wtcnt_recalc();
 	return ((m_wtcsr | 0x18) << 8) | (m_wtcnt & 0xff);
 }
 
-uint16_t sh2_device::rstcsr_r()
+uint16_t sh2_sh7604_device::rstcsr_r()
 {
 	return (m_rstcsr & 0xe0) | 0x1f;
 }
 
-void sh2_device::wtcnt_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::wtcnt_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_wtcw[0]);
 	switch (m_wtcw[0] & 0xff00)
@@ -909,7 +1194,7 @@ void sh2_device::wtcnt_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	}
 }
 
-void sh2_device::rstcsr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::rstcsr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	COMBINE_DATA(&m_wtcw[1]);
 	switch (m_wtcw[1] & 0xff00)
@@ -925,12 +1210,12 @@ void sh2_device::rstcsr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	}
 }
 
-uint16_t sh2_device::fmr_sbycr_r()
+uint16_t sh2_sh7604_device::fmr_sbycr_r()
 {
 	return m_sbycr;
 }
 
-void sh2_device::fmr_sbycr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void sh2_sh7604_device::fmr_sbycr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 {
 	switch (mem_mask)
 	{
@@ -950,12 +1235,12 @@ void sh2_device::fmr_sbycr_w(offs_t offset, uint16_t data, uint16_t mem_mask)
 	}
 }
 
-uint8_t sh2_device::ccr_r()
+uint8_t sh2_sh7604_device::ccr_r()
 {
 	return m_ccr & ~0x30;
 }
 
-void sh2_device::ccr_w(uint8_t data)
+void sh2_sh7604_device::ccr_w(uint8_t data)
 {
 	/*
 	    xx-- ---- Way 0/1
@@ -968,79 +1253,79 @@ void sh2_device::ccr_w(uint8_t data)
 	m_ccr = data;
 }
 
-uint32_t sh2_device::bcr1_r()
+uint32_t sh2_sh7604_device::bcr1_r()
 {
 	return (m_bcr1 & ~0xe008) | (m_is_slave ? 0x8000 : 0);
 }
 
-void sh2_device::bcr1_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::bcr1_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_bcr1);
 }
 
-uint32_t sh2_device::bcr2_r()
+uint32_t sh2_sh7604_device::bcr2_r()
 {
 	return m_bcr2;
 }
 
-void sh2_device::bcr2_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::bcr2_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_bcr2);
 }
 
-uint32_t sh2_device::wcr_r()
+uint32_t sh2_sh7604_device::wcr_r()
 {
 	return m_wcr;
 }
 
-void sh2_device::wcr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::wcr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_wcr);
 }
 
-uint32_t sh2_device::mcr_r()
+uint32_t sh2_sh7604_device::mcr_r()
 {
 	return m_mcr & ~0x103;
 }
 
-void sh2_device::mcr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::mcr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_mcr);
 }
 
-uint32_t sh2_device::rtcsr_r()
+uint32_t sh2_sh7604_device::rtcsr_r()
 {
 	return m_rtcsr & 0xf8;
 }
 
-void sh2_device::rtcsr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::rtcsr_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_rtcsr);
 }
 
-uint32_t sh2_device::rtcnt_r()
+uint32_t sh2_sh7604_device::rtcnt_r()
 {
 	return m_rtcnt & 0xff;
 }
 
-void sh2_device::rtcnt_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::rtcnt_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_rtcnt);
 	m_rtcnt &= 0xff;
 }
 
-uint32_t sh2_device::rtcor_r()
+uint32_t sh2_sh7604_device::rtcor_r()
 {
 	return m_rtcor & 0xff;
 }
 
-void sh2_device::rtcor_w(offs_t offset, uint32_t data, uint32_t mem_mask)
+void sh2_sh7604_device::rtcor_w(offs_t offset, uint32_t data, uint32_t mem_mask)
 {
 	COMBINE_DATA(&m_rtcor);
 	m_rtcor &= 0xff;
 }
 
-void sh2_device::set_frt_input(int state)
+void sh2_sh7604_device::set_frt_input(int state)
 {
 	if(m_frt_input == state) {
 		return;
@@ -1065,7 +1350,7 @@ void sh2_device::set_frt_input(int state)
 	sh2_recalc_irq();
 }
 
-void sh2_device::sh2_recalc_irq()
+void sh2_sh7604_device::sh2_recalc_irq()
 {
 	int irq = 0, vector = -1;
 	int  level;
@@ -1119,135 +1404,4 @@ void sh2_device::sh2_recalc_irq()
 	m_sh2_state->internal_irq_level = irq;
 	m_internal_irq_vector = vector;
 	m_test_irq = 1;
-}
-
-/*
- SH-7021 on-chip device
- */
-
-void sh2a_device::sh7032_dma_exec(int ch)
-{
-	const short dma_word_size[4] = { 0, +1, -1, 0 };
-	uint8_t rs = (m_dma[ch].chcr >> 8) & 0xf; /**< Resource Select bits */
-	if(rs != 0xc) // Auto-Request
-	{
-		logerror("Warning: SH7032 DMA enables non auto-request transfer\n");
-		return;
-	}
-
-	// channel enable & master enable
-	if((m_dma[ch].chcr & 1) == 0 || (m_dmaor & 1) == 0)
-		return;
-
-	printf("%08x %08x %04x\n",m_dma[ch].sar,m_dma[ch].dar,m_dma[ch].chcr);
-	uint8_t dm = (m_dma[ch].chcr >> 14) & 3;  /**< Destination Address Mode bits */
-	uint8_t sm = (m_dma[ch].chcr >> 12) & 3;  /**< Source Address Mode bits */
-	bool ts = (m_dma[ch].chcr & 8);         /**< Transfer Size bit */
-	int src_word_size = dma_word_size[sm] * ((ts == true) ? 2 : 1);
-	int dst_word_size = dma_word_size[dm] * ((ts == true) ? 2 : 1);
-	uint32_t src_addr = m_dma[ch].sar;
-	uint32_t dst_addr = m_dma[ch].dar;
-	uint32_t size_index = m_dma[ch].tcr;
-	if(size_index == 0)
-		size_index = 0x10000;
-
-	if(ts == false)
-		logerror("SH7032: DMA byte mode check\n");
-
-	for(int index = size_index;index>-1;index--)
-	{
-		if(ts == true)
-			m_program->write_word(dst_addr,m_program->read_word(src_addr));
-		else
-			m_program->write_byte(dst_addr,m_program->read_byte(src_addr));
-
-		src_addr += src_word_size;
-		dst_addr += dst_word_size;
-	}
-
-	m_dma[ch].chcr &= ~1; /**< @todo non-instant DMA */
-	printf("%02x %02x %02x %1d\n",sm,dm,rs,ts);
-}
-
-uint32_t sh2a_device::dma_sar0_r()
-{
-	return m_dma[0].sar;
-}
-
-void sh2a_device::dma_sar0_w(offs_t offset, uint32_t data, uint32_t mem_mask)
-{
-	COMBINE_DATA(&m_dma[0].sar);
-}
-
-uint32_t sh2a_device::dma_dar0_r()
-{
-	return m_dma[0].dar;
-}
-
-void sh2a_device::dma_dar0_w(offs_t offset, uint32_t data, uint32_t mem_mask)
-{
-	COMBINE_DATA(&m_dma[0].dar);
-}
-
-uint16_t sh2a_device::dma_tcr0_r()
-{
-	return m_dma[0].tcr;
-}
-
-void sh2a_device::dma_tcr0_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	//printf("%04x\n",data);
-	COMBINE_DATA(&m_dma[0].tcr);
-}
-
-uint16_t sh2a_device::dma_chcr0_r()
-{
-	return m_dma[0].chcr;
-}
-
-void sh2a_device::dma_chcr0_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	//printf("%04x CHCR0\n",data);
-	COMBINE_DATA(&m_dma[0].chcr);
-	sh7032_dma_exec(0);
-}
-
-uint16_t sh2a_device::dmaor_r()
-{
-	return m_dmaor;
-}
-
-void sh2a_device::dmaor_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	COMBINE_DATA(&m_dmaor);
-	sh7032_dma_exec(0);
-}
-
-/*!
-  @brief Dummy debug interface
-  */
-uint16_t sh1_device::sh7032_r(offs_t offset)
-{
-	return m_sh7032_regs[offset];
-}
-
-/*!
-  @brief Dummy debug interface
- */
-void sh1_device::sh7032_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	COMBINE_DATA(&m_sh7032_regs[offset]);
-}
-
-uint16_t sh2a_device::sh7021_r(offs_t offset)
-{
-	return m_sh7021_regs[offset];
-}
-
-/*!
-  @brief Dummy debug interface
- */
-void sh2a_device::sh7021_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	COMBINE_DATA(&m_sh7021_regs[offset]);
 }
