+++ src/devices/cpu/sh/sh7604.h
@@ -1,0 +1,283 @@
+// license:BSD-3-Clause
+// copyright-holders:Juergen Buchmueller, R. Belmont, Angelo Salese, ...
+
+#ifndef MAME_CPU_SH_SH7604_H
+#define MAME_CPU_SH_SH7604_H
+
+#pragma once
+
+#include "sh2.h"
+
+#define SH2_DMA_KLUDGE_CB(name)  int name(uint32_t src, uint32_t dst, uint32_t data, int size)
+#define SH2_DMA_FIFO_DATA_AVAILABLE_CB(name)  int name(uint32_t src, uint32_t dst, uint32_t data, int size)
+#define SH2_FTCSR_READ_CB(name)  void name(uint32_t data)
+
+
+class sh2_sh7604_device : public sh2_device
+{
+public:
+	typedef device_delegate<int (uint32_t src, uint32_t dst, uint32_t data, int size)> dma_kludge_delegate;
+	typedef device_delegate<int (uint32_t src, uint32_t dst, uint32_t data, int size)> dma_fifo_data_available_delegate;
+	typedef device_delegate<void (uint32_t data)> ftcsr_read_delegate;
+
+	// construction/destruction
+	sh2_sh7604_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock);
+
+	void set_is_slave(int slave) { m_is_slave = slave; }
+
+	template <typename... T> void set_dma_kludge_callback(T &&... args) { m_dma_kludge_cb.set(std::forward<T>(args)...); }
+
+	template <typename... T> void set_dma_fifo_data_available_callback(T &&... args) { m_dma_fifo_data_available_cb.set(std::forward<T>(args)...); }
+
+	template <typename... T> void set_ftcsr_read_callback(T &&... args) { m_ftcsr_read_cb.set(std::forward<T>(args)...); }
+
+	void sh2_notify_dma_data_available();
+
+protected:
+	sh2_sh7604_device(const machine_config &mconfig, device_type type, const char *tag, device_t *owner, uint32_t clock, int cpu_type,address_map_constructor internal_map, int addrlines);
+
+	// device-level overrides
+	virtual void device_start() override;
+	virtual void device_reset() override;
+
+	virtual void sh2_exception(const char *message, int irqline) override;
+
+	uint32_t  m_test_irq = 0;
+	int     m_internal_irq_vector = 0;
+
+private:
+	enum
+	{
+		ICF  = 0x80,
+		OCFA = 0x08,
+		OCFB = 0x04,
+		OVF  = 0x02,
+		CCLRA = 0x01
+	};
+
+	void sh7604_map(address_map &map);
+
+	uint32_t sh2_internal_a5();
+
+	// SCI
+	uint8_t smr_r();
+	void smr_w(uint8_t data);
+	uint8_t brr_r();
+	void brr_w(uint8_t data);
+	uint8_t scr_r();
+	void scr_w(uint8_t data);
+	uint8_t tdr_r();
+	void tdr_w(uint8_t data);
+	uint8_t ssr_r();
+	void ssr_w(uint8_t data);
+	uint8_t rdr_r();
+
+	// FRT / FRC
+	uint8_t tier_r();
+	void tier_w(uint8_t data);
+	uint16_t frc_r();
+	void frc_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint8_t ftcsr_r();
+	void ftcsr_w(uint8_t data);
+	uint16_t ocra_b_r();
+	void ocra_b_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint8_t frc_tcr_r();
+	void frc_tcr_w(uint8_t data);
+	uint8_t tocr_r();
+	void tocr_w(uint8_t data);
+	uint16_t frc_icr_r();
+
+	// INTC
+	uint16_t ipra_r();
+	void ipra_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t iprb_r();
+	void iprb_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t vcra_r();
+	void vcra_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t vcrb_r();
+	void vcrb_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t vcrc_r();
+	void vcrc_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t vcrd_r();
+	void vcrd_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t vcrwdt_r();
+	void vcrwdt_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint32_t vcrdiv_r();
+	void vcrdiv_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint16_t intc_icr_r();
+	void intc_icr_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+
+	// DIVU
+	uint32_t dvsr_r();
+	void dvsr_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t dvdnt_r();
+	void dvdnt_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t dvdnth_r();
+	void dvdnth_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t dvdntl_r();
+	void dvdntl_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+
+	uint32_t dvcr_r();
+	void dvcr_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+
+	// DMAC
+	template <int Channel> uint32_t vcrdma_r()
+	{
+		return m_vcrdma[Channel] & 0x7f;
+	}
+
+	template <int Channel> void vcrdma_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0)
+	{
+		COMBINE_DATA(&m_vcrdma[Channel]);
+		m_irq_vector.dmac[Channel] = m_vcrdma[Channel] & 0x7f;
+		sh2_recalc_irq();
+	}
+
+	template <int Channel> uint8_t drcr_r() { return m_dmac[Channel].drcr & 3; }
+	template <int Channel> void drcr_w(uint8_t data) { m_dmac[Channel].drcr = data & 3; sh2_recalc_irq(); }
+	template <int Channel> uint32_t sar_r() { return m_dmac[Channel].sar; }
+	template <int Channel> void sar_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0) { COMBINE_DATA(&m_dmac[Channel].sar); }
+	template <int Channel> uint32_t dar_r() { return m_dmac[Channel].dar; }
+	template <int Channel> void dar_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0) { COMBINE_DATA(&m_dmac[Channel].dar); }
+	template <int Channel> uint32_t dmac_tcr_r() { return m_dmac[Channel].tcr; }
+	template <int Channel> void dmac_tcr_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0) { COMBINE_DATA(&m_dmac[Channel].tcr); m_dmac[Channel].tcr &= 0xffffff; }
+	template <int Channel> uint32_t chcr_r() { return m_dmac[Channel].chcr; }
+	template <int Channel> void chcr_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0)
+	{
+		uint32_t old;
+		old = m_dmac[Channel].chcr;
+		COMBINE_DATA(&m_dmac[Channel].chcr);
+		m_dmac[Channel].chcr = (data & ~2) | (old & m_dmac[Channel].chcr & 2);
+		sh2_dmac_check(Channel);
+	}
+	uint32_t dmaor_r() { return m_dmaor & 0xf; }
+	void dmaor_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0)
+	{
+		if(ACCESSING_BITS_0_7)
+		{
+			uint8_t old;
+			old = m_dmaor & 0xf;
+			m_dmaor = (data & ~6) | (old & m_dmaor & 6); // TODO: should this be old & data & 6? bug?
+			sh2_dmac_check(0);
+			sh2_dmac_check(1);
+		}
+	}
+
+	// WTC
+	uint16_t wtcnt_r();
+	void wtcnt_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint16_t rstcsr_r();
+	void rstcsr_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+
+	// misc
+	uint16_t fmr_sbycr_r();
+	void fmr_sbycr_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	uint8_t ccr_r();
+	void ccr_w(uint8_t data);
+
+	// BSC
+	uint32_t bcr1_r();
+	void bcr1_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t bcr2_r();
+	void bcr2_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t wcr_r();
+	void wcr_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t mcr_r();
+	void mcr_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t rtcsr_r();
+	void rtcsr_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t rtcor_r();
+	void rtcor_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+	uint32_t rtcnt_r();
+	void rtcnt_w(offs_t offset, uint32_t data, uint32_t mem_mask = ~0);
+
+	virtual void set_frt_input(int state) override;
+
+	// SCI
+	uint8_t m_smr = 0, m_brr = 0, m_scr = 0, m_tdr = 0, m_ssr = 0;
+	// FRT / FRC
+	uint8_t m_tier = 0, m_ftcsr = 0, m_frc_tcr = 0, m_tocr = 0;
+	uint16_t m_frc = 0;
+	uint16_t m_ocra = 0, m_ocrb = 0, m_frc_icr = 0;
+	// INTC
+	struct {
+		uint8_t frc = 0;
+		uint8_t sci = 0;
+		uint8_t divu = 0;
+		uint8_t dmac = 0;
+		uint8_t wdt = 0;
+	} m_irq_level;
+	struct {
+		uint8_t fic = 0;
+		uint8_t foc = 0;
+		uint8_t fov = 0;
+		uint8_t divu = 0;
+		uint8_t dmac[2] = { 0, 0 };
+	} m_irq_vector;
+	uint16_t m_ipra = 0, m_iprb = 0;
+	uint16_t m_vcra = 0, m_vcrb = 0, m_vcrc = 0, m_vcrd = 0, m_vcrwdt = 0, m_vcrdiv = 0, m_intc_icr = 0, m_vcrdma[2] = { 0, 0, };
+	bool m_vecmd = false, m_nmie = false;
+
+	// DIVU
+	bool m_divu_ovf = false, m_divu_ovfie = false;
+	uint32_t m_dvsr = 0, m_dvdntl = 0, m_dvdnth = 0;
+
+	// WTC
+	uint8_t m_wtcnt = 0, m_wtcsr = 0;
+	uint8_t m_rstcsr = 0;
+	uint16_t m_wtcw[2] = { 0, 0 };
+
+	// DMAC
+	struct {
+		uint8_t drcr = 0;
+		uint32_t sar = 0;
+		uint32_t dar = 0;
+		uint32_t tcr = 0;
+		uint32_t chcr = 0;
+	} m_dmac[2];
+	uint8_t m_dmaor = 0;
+
+	// misc
+	uint8_t m_sbycr = 0, m_ccr = 0;
+
+	// BSC
+	uint32_t m_bcr1 = 0, m_bcr2 = 0, m_wcr = 0, m_mcr = 0, m_rtcsr = 0, m_rtcor = 0, m_rtcnt = 0;
+
+	uint64_t  m_frc_base = 0;
+
+	int     m_frt_input = 0;
+
+	emu_timer *m_timer = nullptr;
+	emu_timer *m_wdtimer = nullptr;
+	emu_timer *m_dma_current_active_timer[2] { nullptr, nullptr };
+	int     m_dma_timer_active[2] = { 0, 0 };
+	uint8_t  m_dma_irq[2] = { 0, 0 };
+
+	int m_active_dma_incs[2] = { 0, 0 };
+	int m_active_dma_incd[2] = { 0, 0 };
+	int m_active_dma_size[2] = { 0, 0 };
+	int m_active_dma_steal[2] = { 0, 0 };
+	uint32_t m_active_dma_src[2] = { 0, 0 };
+	uint32_t m_active_dma_dst[2] = { 0, 0 };
+	uint32_t m_active_dma_count[2] = { 0, 0 };
+
+	int     m_is_slave = 0;
+	dma_kludge_delegate              m_dma_kludge_cb;
+	dma_fifo_data_available_delegate m_dma_fifo_data_available_cb;
+	ftcsr_read_delegate              m_ftcsr_read_cb;
+
+	TIMER_CALLBACK_MEMBER( sh2_timer_callback );
+	TIMER_CALLBACK_MEMBER( sh2_wdtimer_callback );
+	TIMER_CALLBACK_MEMBER( sh2_dma_current_active_callback );
+	void sh2_timer_resync();
+	void sh2_timer_activate();
+	void sh2_wtcnt_recalc();
+	void sh2_wdt_activate();
+	void sh2_do_dma(int dmach);
+	void sh2_dmac_check(int dma);
+	void sh2_recalc_irq();
+};
+
+DECLARE_DEVICE_TYPE(SH2_SH7604, sh2_sh7604_device)
+
+#endif // MAME_CPU_SH_SH7604_H
