+++ src/devices/cpu/sh/sh7604_bus.h
@@ -6,8 +6,8 @@
 
 ***************************************************************************/
 
-#ifndef MAME_CPU_SH2_SH7604_BUS_H
-#define MAME_CPU_SH2_SH7604_BUS_H
+#ifndef MAME_CPU_SH_SH7604_BUS_H
+#define MAME_CPU_SH_SH7604_BUS_H
 
 #pragma once
 
@@ -67,4 +67,4 @@ private:
 // device type definition
 DECLARE_DEVICE_TYPE(SH7604_BUS, sh7604_bus_device)
 
-#endif // MAME_CPU_SH2_SH7604_BUS_H
+#endif // MAME_CPU_SH_SH7604_BUS_H
