+++ src/devices/cpu/sh/sh_fe.cpp
@@ -10,7 +10,6 @@
 
 #include "emu.h"
 #include "sh2.h"
-#include "sh2comn.h"
 #include "cpu/drcfe.h"
