+++ src/mame/capcom/cps3.cpp
@@ -582,8 +582,6 @@ Hardware registers info
 
 #include "emu.h"
 #include "cdrom.h"
-#include "cpu/sh/sh2.h"
-#include "machine/intelfsh.h"
 #include "machine/nvram.h"
 #include "cps3.h"
 #include "bus/nscsi/cd.h"
@@ -2491,7 +2489,7 @@ void cps3_state::simm6_128mbit(machine_config &config)
 void cps3_state::cps3(machine_config &config)
 {
 	/* basic machine hardware */
-	SH2(config, m_maincpu, 6250000*4); // external clock is 6.25 Mhz, it sets the internal multiplier to 4x (this should probably be handled in the core..)
+	SH2_SH7604(config, m_maincpu, 6250000*4); // external clock is 6.25 Mhz, it sets the internal multiplier to 4x (this should probably be handled in the core..)
 	m_maincpu->set_addrmap(AS_PROGRAM, &cps3_state::cps3_map);
 	m_maincpu->set_addrmap(AS_OPCODES, &cps3_state::decrypted_opcodes_map);
 	m_maincpu->set_dma_kludge_callback(FUNC(cps3_state::dma_callback));
