+++ src/mame/psikyo/psikyo4.h
@@ -6,7 +6,7 @@
 
 *************************************************************************/
 
-#include "cpu/sh/sh2.h"
+#include "cpu/sh/sh7604.h"
 #include "sound/ymopl.h"
 #include "machine/eepromser.h"
 #include "emupal.h"
@@ -57,7 +57,7 @@ private:
 	double         m_oldbrt[2]{};
 
 	/* devices */
-	required_device<sh2_device> m_maincpu;
+	required_device<sh2_sh7604_device> m_maincpu;
 	required_device<eeprom_serial_93cxx_device> m_eeprom;
 	required_device<gfxdecode_device> m_gfxdecode;
 	required_device_array<palette_device, 2> m_palette;
