+++ src/mame/psikyo/psikyosh.cpp
@@ -335,7 +335,6 @@ ROM Usage
 #include "emu.h"
 #include "psikyosh.h"
 
-#include "cpu/sh/sh2.h"
 #include "machine/eepromser.h"
 #include "machine/watchdog.h"
 #include "sound/ymopl.h"
@@ -838,7 +837,7 @@ void psikyosh_state::machine_start()
 void psikyosh_state::psikyo3v1(machine_config &config)
 {
 	/* basic machine hardware */
-	SH2(config, m_maincpu, MASTER_CLOCK/2);
+	SH2_SH7604(config, m_maincpu, MASTER_CLOCK/2);
 	m_maincpu->set_addrmap(AS_PROGRAM, &psikyosh_state::ps3v1_map);
 	m_maincpu->set_vblank_int("screen", FUNC(psikyosh_state::interrupt));
