+++ src/mame/sega/coolridr.cpp
@@ -283,7 +283,8 @@ to the same bank as defined through A20.
 
 #include "emu.h"
 #include "cpu/m68000/m68000.h"
-#include "cpu/sh/sh2.h"
+#include "cpu/sh/sh7032.h"
+#include "cpu/sh/sh7604.h"
 #include "machine/nvram.h"
 #include "machine/timer.h"
 #include "315_5649.h"
@@ -350,8 +351,8 @@ public:
 	uint32_t m_clipvals[2][3];
 	uint8_t  m_clipblitterMode[2]; // hack
 
-	required_device<sh2_device> m_maincpu;
-	required_device<sh2_device> m_subcpu;
+	required_device<sh2_sh7604_device> m_maincpu;
+	required_device<sh1_sh7032_device> m_subcpu;
 	required_device<cpu_device> m_soundcpu;
 	//required_device<am9517a_device> m_dmac;
 
@@ -3227,14 +3228,14 @@ void coolridr_state::scsp2_to_sh1_irq(int state)
 
 void coolridr_state::coolridr(machine_config &config)
 {
-	SH2(config, m_maincpu, XTAL(28'000'000)); // 28 MHz
+	SH2_SH7604(config, m_maincpu, XTAL(28'000'000)); // 28 MHz
 	m_maincpu->set_addrmap(AS_PROGRAM, &coolridr_state::coolridr_h1_map);
 	TIMER(config, "scantimer").configure_scanline(FUNC(coolridr_state::interrupt_main), "screen", 0, 1);
 
 	M68000(config, m_soundcpu, XTAL(32'000'000)/2); // 16 MHz
 	m_soundcpu->set_addrmap(AS_PROGRAM, &coolridr_state::system_h1_sound_map);
 
-	SH1(config, m_subcpu, XTAL(32'000'000)/2); // SH7032 HD6417032F20!! 16 MHz
+	SH1_SH7032(config, m_subcpu, XTAL(32'000'000)/2); // SH7032 HD6417032F20!! 16 MHz
 	m_subcpu->set_addrmap(AS_PROGRAM, &coolridr_state::coolridr_submap);
 	TIMER(config, "scantimer2").configure_scanline(FUNC(coolridr_state::interrupt_sub), "screen", 0, 1);
