+++ src/mame/sega/stv.cpp
@@ -45,7 +45,7 @@
 
 #include "cpu/m68000/m68000.h"
 #include "cpu/scudsp/scudsp.h"
-#include "cpu/sh/sh2.h"
+#include "cpu/sh/sh7604.h"
 #include "machine/smpc.h"
 #include "machine/stvcd.h"
 #include "sound/scsp.h"
@@ -1113,12 +1113,12 @@ void stv_state::pdr2_output_w(uint8_t data)
 void stv_state::stv(machine_config &config)
 {
 	/* basic machine hardware */
-	SH2(config, m_maincpu, MASTER_CLOCK_352/2); // 28.6364 MHz
+	SH2_SH7604(config, m_maincpu, MASTER_CLOCK_352/2); // 28.6364 MHz
 	m_maincpu->set_addrmap(AS_PROGRAM, &stv_state::stv_mem);
 	m_maincpu->set_is_slave(0);
 	TIMER(config, "scantimer").configure_scanline(FUNC(stv_state::saturn_scanline), "screen", 0, 1);
 
-	SH2(config, m_slave, MASTER_CLOCK_352/2); // 28.6364 MHz
+	SH2_SH7604(config, m_slave, MASTER_CLOCK_352/2); // 28.6364 MHz
 	m_slave->set_addrmap(AS_PROGRAM, &stv_state::stv_mem);
 	m_slave->set_is_slave(1);
 	TIMER(config, "slave_scantimer").configure_scanline(FUNC(stv_state::saturn_slave_scanline), "screen", 0, 1);
